+++ src/Model/MetadataBaseRepo.php
@@ -22,7 +22,7 @@ class MetadataBaseRepo
     const PRIMARY_KEY = 'id';
     const TABLE_ALIAS = 'm';
     const FIND_BY_PRIMARY_KEY_SQL = 'SELECT * FROM __meta__ WHERE id = ? LIMIT 1';
-    const DELETE_BY_PRIMARY_KEY_SQL = 'DELETE FROM __meta__ WHERE id = ?';
+    const DELETE_BY_PRIMARY_KEY_SQL = 'DELETE FROM __meta__ WHERE id = ? LIMIT 1';
     public static $columnNames = array (
       0 => 'id',
       1 => 'name',
@@ -36,7 +36,7 @@ class MetadataBaseRepo
     public static $mixinClasses = array (
     );
     protected $table = '__meta__';
-    protected $findStm;
+    protected $loadStm;
     protected $deleteStm;
     public static function getSchema()
     {
@@ -46,13 +46,13 @@ class MetadataBaseRepo
         }
         return $schema = new \LazyRecord\Model\MetadataSchemaProxy;
     }
-    public function findByPrimaryKey($pkId)
+    public function loadByPrimaryKey($pkId)
     {
-        if (!$this->findStm) {
-           $this->findStm = $this->read->prepare(self::FIND_BY_PRIMARY_KEY_SQL);
-           $this->findStm->setFetchMode(PDO::FETCH_CLASS, 'LazyRecord\Model\Metadata');
+        if (!$this->loadStm) {
+           $this->loadStm = $this->read->prepare(self::FIND_BY_PRIMARY_KEY_SQL);
+           $this->loadStm->setFetchMode(PDO::FETCH_CLASS, 'LazyRecord\Model\Metadata');
         }
-        return static::_stmFetch($this->findStm, [$pkId]);
+        return static::_stmFetch($this->loadStm, [$pkId]);
     }
     public function deleteByPrimaryKey($pkId)
     {