+++ src/Schema/Factory/BaseModelClassFactory.php
@@ -21,9 +21,6 @@ use CodeGen\Statement\RequireOnceStatement;
 use CodeGen\Expr\ConcatExpr;
 use CodeGen\Raw;
 
-use LazyRecord\Schema\CodeGenSettingsParser;
-use LazyRecord\Schema\AnnotatedBlock;
-use LazyRecord\Schema\MethodBlockParser;
 
 /**
  * Base Model class generator.
@@ -110,20 +107,6 @@ class BaseModelClassFactory
                 $cTemplate->useTrait($traitClass);
             }
         }
-
-        // parse codegen settings from schema doc comment string
-        $schemaReflection = new ReflectionClass($schema);
-        $schemaDocComment = $schemaReflection->getDocComment();
-
-        $codegenSettings = CodeGenSettingsParser::parse($schemaDocComment);
-        /*
-        if (!empty($codegenSettings)) {
-            $reflectionModel = new ReflectionClass('LazyRecord\\BaseModel');
-            $createMethod = $reflectionModel->getMethod('create');
-            $elements = MethodBlockParser::parseElements($createMethod, 'codegenBlock');
-            $cTemplate->addMethod('public', 'create', ['array $args', 'array $options = array()'], AnnotatedBlock::apply($elements, $codegenSettings));
-        }
-        */
 
         $cTemplate->addStaticMethod('public', 'createRepo', ['$write', '$read'], function() use ($schema) {
             return "return new \\{$schema->getBaseRepoClass()}(\$write, \$read);";